Move BankAccount to Payments namespace
