+++ lib/suma/payment/funding_transaction/increase_ach_strategy.rb
@@ -9,7 +9,7 @@ class Suma::Payment::FundingTransaction::IncreaseAchStrategy <
   include Suma::Payment::FundingTransaction::Strategy
 
   one_to_one :funding_transaction, class: "Suma::Payment::FundingTransaction"
-  many_to_one :originating_bank_account, class: "Suma::BankAccount"
+  many_to_one :originating_bank_account, class: "Suma::Payment::BankAccount"
 
   def originating_instrument
     return self.originating_bank_account
